--- components/libc/compilers/armlibc/unistd.c
@@ -1,25 +1,0 @@
-/*
- * Copyright (c) 2006-2018, RT-Thread Development Team
- *
- * SPDX-License-Identifier: Apache-2.0
- *
- * Change Logs:
- * Date           Author       Notes
- * 2020-09-01     Meco Man     First Version
- */
-
-#include <termios.h>
-#include <unistd.h>
-
-#ifdef RT_USING_POSIX_TERMIOS
-int isatty(int fd)
-{
-    struct termios ts;
-    return(tcgetattr(fd,&ts) != -1);/*true if no error (is a tty)*/
-}
-#endif
-
-char *ttyname(int fd)
-{
-    return "/dev/tty"; /*TODO: need to add more specific*/
-}
